[archimedes] Refactor archimedes
This code provides some refactorings for the archimedes class. It joins the
methods `export_by_id` and `export_by_title` to a single one `export`. Furthermore,
it renames the param `search` to `find` to avoid misunderstandings between the
the kibana search objects and the action of finding files in the folders containing
kibana objects.

The bin/archimedes has changed accordingly.