[archimedes] Bump to version 0.9.0
This code updates archimedes to be able to work
with Kibana 6.8 API.

The method `fetch_objs` in the class `saved_objects`
has been renamed to `find` and aligned to the endpoint
`_find` of the SavedObjects API. This change was
needed since the previous call to the API wasn't
working anymore.

Tests have been updated accordingly.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>