[thales] Derive menu.yaml from metadashboard and projectname
This code proposes a script (thales.py) to generate a menu.yaml
from a metadashboard and projectname included in a .kibana index.
This script is useful to migrate existing top menu to a format
that can be processed by Pythagoras.

Thales can be executed as follows:
```
thales https://admin:admin@localhost:9200 .kibana ./menu.yaml
```

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>